Revert "Cleaning up minimum viable implementation"
