+++ README.md
@@ -4,11 +4,9 @@ Store and compress your train of thought. This project combines [OpenAI Whisper]
 
 # Installation
 
-1. Tested with Python 3.10. Using a Conda environment is recommended: `conda create -n mindbank python=3.10`
+1. Tested with Python 3.10 on OSX. Using a Conda environment is recommended: `conda create -n mindbank python=3.10`
 
-1. `cd` into the repo's root directory (`/mindbank/`) and install python packages: `pip install -r requirements.txt`
-
-1. Install the package: `pip install .`
+1. Install whisper: `pip install -U openai-whisper`
 
 1. Install ffmpeg:
 
@@ -29,8 +27,12 @@ choco install ffmpeg
 scoop install ffmpeg
 ```
 
+4. Install pyaudio: `pip install pyaudio`
+
+5. Install OpenAi: `pip install openai`
+
+6. Install keyboard: `pip install keyboard`
+
 # Usage
 
-First, generate an OpenAI token: https://platform.openai.com/
-
-`python mindbank --token <path_to_api_token>`
+`cd` into `src/mindbank`, run `python mindbank`, and follow the terminal prompts.
